requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-model-utils==4.0.0
drf-yasg
pymysql
django-livesync
requests
psycopg2-binary==2.8.6